+++ src/main/java/edu/cnm/deepdive/join_me_backend/controller/SquareController.java
@@ -55,7 +55,7 @@ public class SquareController {
   public ResponseEntity<List<Square>> post() {
     List<Square> grid = new LinkedList<>();
     Square sq1 = new Square();
-    sq1.setId(100001);
+    sq1.setId(100001L);
     sq1.setLatitudeLowerBound(35.0858000);
     sq1.setLatitudeUpperBound(35.0859000);
     sq1.setLongitudeLowerBound(-106.6497000);
@@ -63,7 +63,7 @@ public class SquareController {
     grid.add(sq1);
 
     Square sq2 = new Square();
-    sq2.setId(200002);
+    sq2.setId(200002L);
     sq2.setLatitudeLowerBound(35.0858000);
     sq2.setLatitudeUpperBound(35.0859000);
     sq2.setLongitudeLowerBound(-106.6496000);
@@ -71,7 +71,7 @@ public class SquareController {
     grid.add(sq2);
 
     Square sq3 = new Square();
-    sq3.setId(300003);
+    sq3.setId(300003L);
     sq3.setLatitudeLowerBound(35.0858000);
     sq3.setLatitudeUpperBound(35.0859000);
     sq3.setLongitudeLowerBound(-106.6495000);
@@ -79,7 +79,7 @@ public class SquareController {
     grid.add(sq3);
 
     Square sq4 = new Square();
-    sq4.setId(400004);
+    sq4.setId(400004L);
     sq4.setLatitudeLowerBound(35.0858000);
     sq4.setLatitudeUpperBound(35.0859000);
     sq4.setLongitudeLowerBound(-106.6494000);
@@ -87,7 +87,7 @@ public class SquareController {
     grid.add(sq4);
 
     Square sq5 = new Square();
-    sq5.setId(500005);
+    sq5.setId(500005L);
     sq5.setLatitudeLowerBound(35.0859000);
     sq5.setLatitudeUpperBound(35.0860000);
     sq5.setLongitudeLowerBound(-106.6497000);
@@ -95,7 +95,7 @@ public class SquareController {
     grid.add(sq5);
 
     Square sq6 = new Square();
-    sq6.setId(600006);
+    sq6.setId(600006L);
     sq6.setLatitudeLowerBound(35.0859000);
     sq6.setLatitudeUpperBound(35.0860000);
     sq6.setLongitudeLowerBound(-106.6496000);
@@ -103,7 +103,7 @@ public class SquareController {
     grid.add(sq6);
 
     Square sq7 = new Square();
-    sq7.setId(700007);
+    sq7.setId(700007L);
     sq7.setLatitudeLowerBound(35.0859000);
     sq7.setLatitudeUpperBound(35.0860000);
     sq7.setLongitudeLowerBound(-106.6495000);
@@ -111,7 +111,7 @@ public class SquareController {
     grid.add(sq7);
 
     Square sq8 = new Square();
-    sq8.setId(800008);
+    sq8.setId(800008L);
     sq8.setLatitudeLowerBound(35.0859000);
     sq8.setLatitudeUpperBound(35.0860000);
     sq8.setLongitudeLowerBound(-106.6494000);
@@ -119,7 +119,7 @@ public class SquareController {
     grid.add(sq8);
 
     Square sq9 = new Square();
-    sq9.setId(900009);
+    sq9.setId(900009L);
     sq9.setLatitudeLowerBound(35.0860000);
     sq9.setLatitudeUpperBound(35.0861000);
     sq9.setLongitudeLowerBound(-106.6497000);
@@ -127,7 +127,7 @@ public class SquareController {
     grid.add(sq9);
 
     Square sq10 = new Square();
-    sq10.setId(10000010);
+    sq10.setId(10000010L);
     sq10.setLatitudeLowerBound(35.0860000);
     sq10.setLatitudeUpperBound(35.0861000);
     sq10.setLongitudeLowerBound(-106.6496000);
@@ -135,7 +135,7 @@ public class SquareController {
     grid.add(sq10);
 
     Square sq11 = new Square();
-    sq11.setId(11000011);
+    sq11.setId(11000011L);
     sq11.setLatitudeLowerBound(35.0860000);
     sq11.setLatitudeUpperBound(35.0861000);
     sq11.setLongitudeLowerBound(-106.6495000);
@@ -143,7 +143,7 @@ public class SquareController {
     grid.add(sq11);
 
     Square sq12 = new Square();
-    sq12.setId(12000012);
+    sq12.setId(12000012L);
     sq12.setLatitudeLowerBound(35.0860000);
     sq12.setLatitudeUpperBound(35.0861000);
     sq12.setLongitudeLowerBound(-106.6494000);
@@ -151,7 +151,7 @@ public class SquareController {
     grid.add(sq12);
 
     Square sq13 = new Square();
-    sq13.setId(13000013);
+    sq13.setId(13000013L);
     sq13.setLatitudeLowerBound(35.0861000);
     sq13.setLatitudeUpperBound(35.0862000);
     sq13.setLongitudeLowerBound(-106.6497000);
@@ -159,7 +159,7 @@ public class SquareController {
     grid.add(sq13);
 
     Square sq14 = new Square();
-    sq14.setId(14000014);
+    sq14.setId(14000014L);
     sq14.setLatitudeLowerBound(35.0861000);
     sq14.setLatitudeUpperBound(35.0862000);
     sq14.setLongitudeLowerBound(-106.6496000);
@@ -167,7 +167,7 @@ public class SquareController {
     grid.add(sq14);
 
     Square sq15 = new Square();
-    sq15.setId(15000015);
+    sq15.setId(15000015L);
     sq15.setLatitudeLowerBound(35.0861000);
     sq15.setLatitudeUpperBound(35.0862000);
     sq15.setLongitudeLowerBound(-106.6495000);
@@ -175,7 +175,7 @@ public class SquareController {
     grid.add(sq15);
 
     Square sq16 = new Square();
-    sq16.setId(16000016);
+    sq16.setId(16000016L);
     sq16.setLatitudeLowerBound(35.0861000);
     sq16.setLatitudeUpperBound(35.0862000);
     sq16.setLongitudeLowerBound(-106.6494000);
@@ -187,14 +187,14 @@ public class SquareController {
   }
 
   @GetMapping(value = "{squareId}", produces = MediaType.APPLICATION_JSON_VALUE)
-  public Square get(@PathVariable("squareId") int squareId) {
+  public Square get(@PathVariable("squareId") long squareId) {
     return squareRepository.findById(squareId).get();
   }
 
   @Transactional
   @DeleteMapping(value = "{squareId}")
   @ResponseStatus(HttpStatus.NO_CONTENT)
-  public void delete(@PathVariable("squareId") int squareId) {
+  public void delete(@PathVariable("squareId") long squareId) {
     Square square = get(squareId);
     List<Vertex> vertices = square.getVertices();
     for (Vertex vertex : vertices) {
@@ -206,13 +206,13 @@ public class SquareController {
 
   // TODONE Add controller method to return list of Student instances for a specified projectId.
   @GetMapping("{squareId}/vertices")
-  public List<Vertex> vertexList (@PathVariable("squareId") int squareId) {
+  public List<Vertex> vertexList (@PathVariable("squareId") long squareId) {
     return get(squareId).getVertices();
   }
 
   @PostMapping(value = "{squareId}/vertices", consumes = MediaType.APPLICATION_JSON_VALUE,
       produces = MediaType.APPLICATION_JSON_VALUE)
-  public ResponseEntity<Square> postVertex(@PathVariable("squareId") int squareId,
+  public ResponseEntity<Square> postVertex(@PathVariable("squareId") long squareId,
       @RequestBody Vertex partialVertex) {
     Square square = get(squareId);
     Vertex vertex = vertexRepository.findById(partialVertex.getId()).get();
@@ -223,8 +223,8 @@ public class SquareController {
 
   @DeleteMapping(value = "{squareId}/vertices/{vertexId}")
   @ResponseStatus(HttpStatus.NO_CONTENT)
-  public void deleteVertex(@PathVariable("squareId") int squareId,
-      @PathVariable("vertexId") int vertexId) {
+  public void deleteVertex(@PathVariable("squareId") long squareId,
+      @PathVariable("vertexId") long vertexId) {
     Square square = get(squareId);
     Vertex vertex = vertexRepository.findById(vertexId).get();
     if (square.getVertices().remove(vertex)){
